Split horizontal and vertical logic into separate files
